+++ miniRT/src/geometry/sphere.c
@@ -7,15 +7,15 @@
 
 t_intersection *intersect_sphere(t_ray *ray, t_sphere *sphere) {
     t_vector oc = vector_subtract(ray->origin, sphere->center);
-    double a = vector_dot_product(ray->direction, ray->direction);
-    double b = 2.0 * vector_dot_product(oc, ray->direction);
-    double c = vector_dot_product(oc, oc) - (sphere->diameter * sphere->diameter / 4.0);
-    double discriminant = b * b - 4 * a * c;
+    float a = vector_dot_product(ray->direction, ray->direction);
+    float b = 2.0 * vector_dot_product(oc, ray->direction);
+    float c = vector_dot_product(oc, oc) - (sphere->diameter * sphere->diameter / 4.0);
+    float discriminant = b * b - 4 * a * c;
 
     if (discriminant < 0)
         return NULL;
 
-    double t = (-b - sqrt(discriminant)) / (2.0 * a);
+    float t = (-b - sqrt(discriminant)) / (2.0 * a);
     if (t < 0) {
         t = (-b + sqrt(discriminant)) / (2.0 * a);
         if (t < 0)
