+++ miniRT/src/geometry/vector.c
@@ -5,7 +5,7 @@
 
 #include "../../include/miniRT.h"
 
-t_vector vector_create(double x, double y, double z) {
+t_vector vector_create(float x, float y, float z) {
     return (t_vector){x, y, z};
 }
 
@@ -17,7 +17,7 @@ t_vector vector_subtract(t_vector v1, t_vector v2) {
     return vector_create(v1.x - v2.x, v1.y - v2.y, v1.z - v2.z);
 }
 
-double vector_dot_product(t_vector v1, t_vector v2) {
+float vector_dot_product(t_vector v1, t_vector v2) {
     return (v1.x * v2.x) + (v1.y * v2.y) + (v1.z * v2.z);
 }
 
@@ -29,20 +29,20 @@ t_vector vector_cross_product(t_vector v1, t_vector v2) {
     );
 }
 
-double vector_magnitude(t_vector v) {
+float vector_magnitude(t_vector v) {
     return sqrt(vector_dot_product(v, v));
 }
 
 t_vector vector_normalize(t_vector v) {
-    double mag = vector_magnitude(v);
+    float mag = vector_magnitude(v);
     if (mag == 0.0) return v;  // Avoid division by zero
     return vector_create(v.x / mag, v.y / mag, v.z / mag);
 }
 
-t_vector vector_multiply(t_vector v, double scalar) {
+t_vector vector_multiply(t_vector v, float scalar) {
     return vector_create(v.x * scalar, v.y * scalar, v.z * scalar);
 }
 
-double vector_distance(t_vector v1, t_vector v2) {
+float vector_distance(t_vector v1, t_vector v2) {
     return vector_magnitude(vector_subtract(v2, v1));
 }
